requirements.txt
aiohttp==0.21.6
cffi==1.7.0
chardet==2.3.0
cleverbot==1.0.2
discord.py==0.10.0
future==0.15.2
pycparser==2.14
PyNaCl==1.0.1
redis==2.10.5
requests==2.10.0
six==1.10.0
websockets==3.1
youtube-dl==2016.7.3
